feat: visited clauses statistic
