+++ src/lib/components/StatisticsComponent.svelte
@@ -1,7 +1,5 @@
 <script lang="ts">
 	import type { Trail } from '$lib/entities/Trail.svelte.ts';
-	import { getClausePool } from '$lib/states/problem.svelte.ts';
-	import { getSolverMachine } from '$lib/states/solver-machine.svelte.ts';
 	import {
 		getNoConflicts,
 		getNoDecisions,
@@ -19,9 +17,7 @@
 			return latestTrail.getDL();
 		} else return 0;
 	});
-	const clausesLeft: number = $derived(getClausePool().leftToSatisfy());
 	const visitedClauses: number = $derived(getVisitedClauses());
-	const unsat: boolean = $derived(getSolverMachine().onUnsatState());
 </script>
 
 <statistics>
@@ -41,16 +37,11 @@
 		<span>UPs:</span>
 		<span class="statistic-value">{unitPropagations}</span>
 	</div>
+
 	<div class="metric">
-		Clauses left:
-		<span class="statistic-value">{clausesLeft}</span>
+		Visited Clauses:
+		<span class="statistic-value">{visitedClauses}</span>
 	</div>
-	{#if unsat}
-		<div class="metric">
-			Minimum Clauses:
-			<span class="statistic-value">{visitedClauses}</span>
-		</div>
-	{/if}
 </statistics>
 
 <style>
