+++ src/lib/solvers/cdcl/cdcl-domain.svelte.ts
@@ -7,8 +7,6 @@ import Clause, {
 import type ClausePool from '$lib/entities/ClausePool.svelte.ts';
 import { ConflictAnalysis, type VirtualResolution } from '$lib/entities/ConflictAnalysis.svelte.ts';
 import {
-	type OccurrenceList,
-	type PreprocessingList,
 	type VisitingOccurrenceList
 } from '$lib/entities/OccurrenceList.svelte.ts';
 import type { Trail } from '$lib/entities/Trail.svelte.ts';
@@ -17,6 +15,7 @@ import type { VariablePool } from '$lib/entities/VariablePool.svelte.ts';
 import {
 	atLevelZero,
 	clauseEvaluation,
+	getNextClause,
 	allAssigned as solverAllAssigned,
 	complementaryOccurrences as solverComplementaryOccurrences,
 	decide as solverDecide,
@@ -26,7 +25,6 @@ import {
 import { getConflictAnalysis, setConflictAnalysis } from '$lib/states/conflict-anlysis.svelte.ts';
 import {
 	getClausePool,
-	getCurrentOccurrences,
 	getOccurrenceListQueue,
 	getOccurrencesTableMapping,
 	getProblemStore,
@@ -167,14 +165,7 @@ export const traversedOccurrenceList: CDCL_TRAVERSED_OCCURRENCE_LIST_FUN = (
 export type CDCL_NEXT_OCCURRENCE_FUN = () => CRef;
 
 export const nextClause: CDCL_NEXT_OCCURRENCE_FUN = () => {
-	const visitingOccurrences: VisitingOccurrenceList = getCurrentOccurrences();
-	const unwrappedOccurrences: NonNullable<PreprocessingList | OccurrenceList> =
-		unwrapEither(visitingOccurrences);
-	if (unwrappedOccurrences.isEmpty()) {
-		logFatal('The occurrence list is empty');
-	} else {
-		return unwrappedOccurrences.next();
-	}
+	return getNextClause();
 };
 
 export type CDCL_CONFLICT_DETECTION_FUN = (cRef: CRef) => boolean;
