+++ src/lib/solvers/shared.svelte.ts
@@ -2,6 +2,11 @@ import type Clause from '$lib/entities/Clause.svelte.ts';
 import type { ClauseEval } from '$lib/entities/Clause.svelte.ts';
 import type ClausePool from '$lib/entities/ClausePool.svelte.ts';
 import Literal from '$lib/entities/Literal.svelte.ts';
+import type {
+	OccurrenceList,
+	PreprocessingList,
+	VisitingOccurrenceList
+} from '$lib/entities/OccurrenceList.svelte.ts';
 import type { EWC } from '$lib/entities/Problem.svelte.ts';
 import { Trail } from '$lib/entities/Trail.svelte.ts';
 import type Variable from '$lib/entities/Variable.svelte.ts';
@@ -11,12 +16,16 @@ import type { VariablePool } from '$lib/entities/VariablePool.svelte.ts';
 import { decisionMadeEventBus, newTrailStackedEventBus } from '$lib/events/events.ts';
 import { getAssignment, type AssignmentEvent } from '$lib/states/assignment.svelte.ts';
 import { isBreakpoint } from '$lib/states/breakpoints.svelte.ts';
-import { getVariablePool } from '$lib/states/problem.svelte.ts';
+import { getCurrentOccurrences, getVariablePool } from '$lib/states/problem.svelte.ts';
 import { getSolverMachine } from '$lib/states/solver-machine.svelte.ts';
-import { increaseNoConflicts, increaseNoUnitPropagations } from '$lib/states/statistics.svelte.ts';
+import {
+	increaseNoConflicts,
+	increaseNoUnitPropagations,
+	increaseNoVisitedClauses
+} from '$lib/states/statistics.svelte.ts';
 import { logBreakpoint, logFatal } from '$lib/states/toasts.svelte.ts';
 import { getLatestTrail, stackTrail } from '$lib/states/trails.svelte.ts';
-import { fromLeft, fromRight, isLeft } from '$lib/types/either.ts';
+import { fromLeft, fromRight, isLeft, unwrapEither } from '$lib/types/either.ts';
 import type { CRef, Lit, Var } from '$lib/types/types.ts';
 import { fromJust, isJust, type Maybe } from '../types/maybe.ts';
 
@@ -221,4 +230,15 @@ export const finalStateControl = (): void => {
 
 export const obtainCRefFromEWC = (watch: EWC): CRef => {
 	return isLeft(watch) ? fromLeft(watch).cRef : fromRight(watch);
+};
+
+export const getNextClause: () => CRef = () => {
+	const visitingOccurrences: VisitingOccurrenceList = getCurrentOccurrences();
+	const unwrappedOccurrences: NonNullable<PreprocessingList | OccurrenceList> =
+		unwrapEither(visitingOccurrences);
+	if (unwrappedOccurrences.isEmpty()) {
+		logFatal('The occurrence list is empty');
+	}
+	increaseNoVisitedClauses();
+	return unwrappedOccurrences.next();
 };
